third_party/WebKit/LayoutTests/http/tests/websocket/handshake-fail-by-maxlength-expected.txt

CONSOLE ERROR: line 27: WebSocket connection to 'ws://localhost:8880/handshake-fail-by-maxlength' failed: Error during WebSocket handshake: net::ERR_INVALID_HTTP_RESPONSE
Connection should fail immediately, rather than succeeding or staying in limbo until timeout, if handshake is longer than 1024 bytes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS onerror() was called
PASS timedOut is false
PASS connected is false
PASS wsOrigin is undefined.
PASS successfullyParsed is true

TEST COMPLETE
